First step towards a ".h library"
moving stuff from fourier_bank.cc to fourier_bank.h
